Edits for writing pruned DB (#1045)
* alphabetize protein group names

* new alphabetize protein groups

* update

* update with test

* update good test

* ok

* comment

* kljh

* Bug fixes for pruned database generation! Now all proteins that can be identified via non-ambigous psms are written to the pruned database, but modifications are only written on proteins where they are localized

* reconcile w master branch

* fix bugs that were caught by unit tests, builds properly

* UnitTest for protein Pruned DB generation if complete and functions

* fix merge stuff

* Revisions based on edits

* Finished revisions with test restructure

* Codecy Changes

* minor codecy fix

* more codecy fixes

* coadacy

* codacy again

* final codacy edit.. i hope
